refactor(pane): move named keys below mux
Make pane input vocabulary independent of backend control mechanics.
