+++ crates/rimz/src/agents/adapters/claude/ask.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Deserializer};
 use serde_json::{Map, Value};
 
 use crate::agents::{AnswerPlanErr, AnswerStep, AskKind, AskReply};
-use crate::mux::NamedKey;
+use crate::pane::keys::NamedKey;
 use crate::transcript::{AskAnswer, AskOption, AskQuestion};
 
 #[derive(Debug, Default, Deserialize)]
