+++ crates/rimz/src/agents/adapters/codex/ask.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 use serde_json::Value;
 
 use crate::agents::{AnswerPlanErr, AnswerStep, AskKind, AskReply};
-use crate::mux::NamedKey;
+use crate::pane::keys::NamedKey;
 use crate::transcript::{AskAnswer, AskOption, AskQuestion};
 
 const REQUEST_USER_INPUT_TOOL: &str = "request_user_input";
