+++ crates/rimz/src/message/send.rs
@@ -6,7 +6,8 @@ use std::time::Duration;
 use crate::Store;
 use crate::agents::AgentState;
 use crate::message::{MessageDraft, Recipient, command_segments};
-use crate::mux::{NamedKey, paste_into_pane, press_pane_key, type_into_pane};
+use crate::mux::{paste_into_pane, press_pane_key, type_into_pane};
+use crate::pane::keys::NamedKey;
 use crate::store::message::{AutoCompact, MessageBody, MessageRecord, MessageSender};
 use crate::store::snapshot::{PaneAgent, SidebarSnapshot};
 use crate::workspace::ResolvedWorkspace;
