+++ crates/rimz/src/pane/keys.rs
